Move to extensions class
